# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Daily Report App API",
    "version": "1.0.0",
    "description": "API for the Daily Report App"
  },
  "paths": {
    "/api/get-gemini-models": {
      "get": {
        "summary": "Get available Gemini models",
        "responses": {
          "200": {
            "description": "A list of Gemini models",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetGeminiModelsResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "null", "nullable": true },
          "error": { "type": "string" },
          "status": { "type": "number" }
        }
      },
      "GeminiModel": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "displayName": { "type": "string" }
        }
      },
      "GetGeminiModelsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "models": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/GeminiModel"
                }
              }
            }
          },
          "status": { "type": "number", "example": 200 }
        }
      },
...
    }
  }
}